Fix render for repeat
- Fix repeat logic
- Allow bindings for properties in JSONSchema by automatically
  decorating definitions with bindable
- Update example
